tests/requirements.txt
# Python dependencies for integration tests
pytest>=7.4.0
msgpack>=1.0.5
pyzmq>=25.1.0
requests>=2.31.0
